handling API error responses

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,5 @@
 using System;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -6,21 +7,36 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivity
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>   //Unit is a type from MediatR that represents a void return type
     {
         public required string Id { get; set; }
     }
 
-    public class Handler(AppDbContext context) : IRequestHandler<Command>
+    public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
     {
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
             var activityToDelete = await context.Activities
-                .FindAsync([request.Id], cancellationToken)
-                    ?? throw new Exception("Activity to delete not found");
+                .FindAsync([request.Id], cancellationToken);
+
+            //?? throw new Exception("Activity to delete not found"); -> not a good practice to throw exceptions for flow control
+
+            if (activityToDelete == null)
+            {
+                return Result<Unit>.Failure("Activity not found", 404);
+            }
 
             context.Activities.Remove(activityToDelete);
-            await context.SaveChangesAsync(cancellationToken);
+            var inserted = await context.SaveChangesAsync(cancellationToken) > 0;
+
+            if (inserted)
+            {
+                return Result<Unit>.Success(Unit.Value);
+            }
+            else
+            {
+                return Result<Unit>.Failure("Deleting activity failed", 400);
+            }
         }
     }
